requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cssselect==1.2.0
cssutils==2.6.0
Django==4.1.1
django-ckeditor==6.5.1
django-default-imagefield==0.5.5
django-embed-video==1.4.7
django-etc==1.4.0
django-guest-user==0.5.2
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-paypal==2.0
django-taggit==3.0.0
httplib2==0.21.0
idna==3.4
lxml==5.1.0
oauthlib==3.2.2
pillow==10.2.0
premailer==3.10.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
yagmail==0.12.231
